book_writer_ai/requirements.txt
streamlit==1.32.0
google-generativeai>=0.4.1,<0.5.0
python-dotenv==1.0.1
sqlalchemy>=2.0.41 